refactor: Split CarouselItem mock data into separate files
- Renamed `announcement_mock.dart` to `announcements_mock.dart` for consistency with other mock data files
- Moved CarouselItem mock data related to announcements into `announcements_mock.dart`
- Created `events_mock.dart` and moved CarouselItem mock data related to events into this new file
- Created `news_mock.dart` and moved CarouselItem mock data related to news into this new file
- Updated imports in relevant files to use the new mock data files

This refactor improves the organization and maintainability of the mock data by splitting it into separate files based on the content type. It makes it easier to manage and update the mock data for each content type independently. The naming convention for the mock data files has also been standardized for better consistency across the project.